+++ adk-examples/c-sharp/examples/ADKConnector.cs
@@ -10,305 +10,10 @@ using Newtonsoft.Json;
 using System.IO;
 using log4net;
 
-namespace ConsoleApp2
+namespace examples
 {
     class ADKConnnector
     {
-        static String url = "https://example.falkonry.ai";
-        static String token = "token";
-        Falkonry falkonry = new Falkonry(url, token);
-        DatastreamRequest ds = new DatastreamRequest();
-        Datasource dataSource = new Datasource();
-        Field field = new Field();
-        Time time = new Time();
-        Signal signal = new Signal();
-
-        private int CheckStatus(System.String trackerId)
-        {
-            for (int i = 0; i < 12; i++)
-            {
-                Tracker tracker = falkonry.GetStatus(trackerId);
-                if (tracker.Status.Equals("FAILED") || tracker.Status.Equals("ERROR"))
-                {
-                    throw new System.Exception(tracker.Message + "\n");
-                }
-                else if (tracker.Status.Equals("SUCCESS") || tracker.Status.Equals("COMPLETED"))
-                {
-                    Console.Write(tracker.Status + "\n");
-                    return 1;
-                }
-                System.Threading.Thread.Sleep(5000);
-            }
-            return 0;
-        }
-
-        public String CreateDataStream()
-        {
-            String datastreamName = "C#_TEST_4";
-            String timeZone = "GMT";
-            String timeIdentifier = "time";
-            String timeFormat = "millis";
-            String PrecisionFormat = "millis";
-            String entityIdentifier = "entity";
-            //String signalIdentifier = "signal";     //for narrow data format
-            //String valueIdentifier = "value";       //for narrow data format
-            //String batchIdentifier = "batch_id";       //for batch-type data
-
-            time.Format = timeFormat;
-            time.Zone = timeZone;
-            time.Identifier = timeIdentifier;
-            ds.TimePrecision = PrecisionFormat;
-            dataSource.Type = "PI";
-            ds.DataSource = dataSource;
-            ds.Name = datastreamName;
-            ds.Field = field;
-            ds.Field.Time = time;
-            ds.Field.Signal = signal;
-            ds.Field.EntityIdentifier = entityIdentifier;
-            //ds.Field.Signal.SignalIdentifier = signalIdentifier;        //for narrow data format
-            //ds.Field.Signal.ValueIdentifier = valueIdentifier;          //for narrow data format
-            //ds.Field.BatchIdentifier = batchIdentifier;                 //for batch-type data
-            Datastream datastream = falkonry.CreateDatastream(ds);
-            Console.Write("New Datastream with id {0} created. \n", datastream.Id);
-            return datastream.Id;
-        }
-
-        public void PostData(String datastreamId, String stream, String fileType, Boolean liveStatus=false)
-        {
-            if (fileType == null || datastreamId == null || stream == null)
-            {
-                throw new ArgumentNullException(nameof(fileType));
-            }
-            Datastream datastream = falkonry.GetDatastream(datastreamId);
-            var options = new SortedDictionary<string, string>();
-            var inputStatus = new InputStatus();
-            options.Add("streaming", liveStatus.ToString());
-            options.Add("hasMoreData", "false");
-            options.Add("timeIdentifier", datastream.Field.Time.Identifier.ToString());
-            options.Add("timeZone", datastream.Field.Time.Zone.ToString());
-            options.Add("timeFormat", datastream.Field.Time.Format.ToString());
-            options.Add("entityIdentifier", datastream.Field.EntityIdentifier.ToString());
-            //options.Add("signalIdentifier", datastream.Field.Signal.SignalIdentifier.ToString());     // for narrow data format
-            //options.Add("valueIdentifier", datastream.Field.Signal.ValueIdentifier.ToString());       // for narrow data format
-            //options.Add("batchIdentifier", datastream.Field.BatchIdentifier.ToString());              // for batch-type data
-            options.Add("fileFormat", fileType);
-            if (!liveStatus)
-            {
-                int i;
-                for (i = 0; i < 3; i++)
-                {
-                    try
-                    {
-                        inputStatus = falkonry.AddInput(datastreamId, stream, options);
-                        if (CheckStatus(inputStatus.Id) == 1)
-                        {
-                            break;
-                        }
-                    }
-                    catch (FalkonryClient.Service.FalkonryException e)
-                    {
-                        log.Error(e.GetBaseException() + "\n");
-                        Console.ReadKey();
-                        Environment.Exit(1);
-                    }
-                    catch (Exception ex)
-                    {
-                       log.WarnFormat(ex.Message + "\n" + "Retry Attempt: {0}", i + 1);
-                    }
-                }
-                if (i == 3)
-                {
-                    log.Error("Cannot add data to the datastream!");
-                    Console.ReadKey();
-                    Environment.Exit(1);
-                }
-            }
-            else
-            {
-                try
-                {
-                    inputStatus = falkonry.AddInput(datastreamId, stream, options);
-                }
-                catch (FalkonryClient.Service.FalkonryException e)
-                {
-                    log.Error(e.GetBaseException() + "\n");
-                }
-            }
-        }
-
-        public void PostDataFromStream(String datastreamId, byte[] stream, String fileType, Boolean liveStatus = false)
-        {
-            if (fileType == null || datastreamId == null || stream == null)
-            {
-                throw new ArgumentNullException(nameof(fileType));
-            }
-            Datastream datastream = falkonry.GetDatastream(datastreamId);
-            var inputStatus = new InputStatus();
-            var options = new SortedDictionary<string, string>();
-            options.Add("streaming", liveStatus.ToString());
-            options.Add("hasMoreData", "false");
-            options.Add("timeIdentifier", datastream.Field.Time.Identifier.ToString());
-            options.Add("timeZone", datastream.Field.Time.Zone.ToString());
-            options.Add("timeFormat", datastream.Field.Time.Format.ToString());
-            options.Add("entityIdentifier", datastream.Field.EntityIdentifier.ToString());
-            //options.Add("signalIdentifier", datastream.Field.Signal.SignalIdentifier.ToString());     // for narrow data format
-            //options.Add("valueIdentifier", datastream.Field.Signal.ValueIdentifier.ToString());       // for narrow data format
-            //options.Add("batchIdentifier", datastream.Field.BatchIdentifier.ToString());              // for batch-type data
-            options.Add("fileFormat", fileType);
-            if (!liveStatus)
-            {
-                int i;
-                for (i = 0; i < 3; i++)
-                {
-                    try
-                    {
-                        inputStatus = falkonry.AddInputStream(datastreamId, stream, options);
-                        if (CheckStatus(inputStatus.Id) == 1)
-                        {
-                            break;
-                        }
-                    }
-                    catch (FalkonryClient.Service.FalkonryException e)
-                    {
-                        log.Error(e.GetBaseException() + "\n");
-                        Console.ReadKey();
-                        Environment.Exit(1);
-                    }
-                    catch (Exception ex)
-                    {
-                        log.WarnFormat(ex.Message + "\n" + "Retry Attempt: {0}", i + 1);
-                    }
-                }
-                if(i == 3)
-                {
-                    log.Error("Cannot add data to the datastream!");
-                    Console.ReadKey();
-                    Environment.Exit(1);
-                }
-            }
-            else
-            {
-                try
-                {
-                    inputStatus = falkonry.AddInputStream(datastreamId, stream, options);
-                }
-                catch (FalkonryClient.Service.FalkonryException e)
-                {
-                    Console.Write(e.GetBaseException() + "\n");
-                }
-            }
-        }
-
-
-        public void PostMoreDataFromStream(String datastreamId, String folderPath, Boolean liveStatus = false)
-        {
-            if (datastreamId == null || folderPath == null)
-            {
-                throw new ArgumentNullException(nameof(datastreamId));
-            }
-            FileAdapter fileAdapter = new FileAdapter();
-            Datastream datastream = falkonry.GetDatastream(datastreamId);
-            var options = new SortedDictionary<string, string>();
-            options.Add("streaming", liveStatus.ToString());
-            options.Add("hasMoreData", "true");
-            options.Add("timeIdentifier", datastream.Field.Time.Identifier.ToString());
-            options.Add("timeZone", datastream.Field.Time.Zone.ToString());
-            options.Add("timeFormat", datastream.Field.Time.Format.ToString());
-            options.Add("entityIdentifier", datastream.Field.EntityIdentifier.ToString());
-            //options.Add("signalIdentifier", datastream.Field.Signal.SignalIdentifier.ToString());     // for narrow data format
-            //options.Add("valueIdentifier", datastream.Field.Signal.ValueIdentifier.ToString());       // for narrow data format
-            //options.Add("batchIdentifier", datastream.Field.BatchIdentifier.ToString());              // for batch-type data
-            options.Add("fileFormat", "csv");
-            var inputStatus = new InputStatus();
-            String[] Files = Directory.GetFiles(folderPath + "\\", "*.csv");
-            var fileCount = Files.Length;
-            foreach(String file in Files)
-            {
-                byte[] stream = fileAdapter.GetStream(file);
-                if(fileCount == 1)
-                {
-                    options["hasMoreData"] = "false";
-                }
-                fileCount--;
-                if (!liveStatus)
-                {
-                    int i;
-                    for (i = 0; i < 3; i++)
-                    {
-                        try
-                        {
-                            inputStatus = falkonry.AddInputStream(datastreamId, stream, options);
-                            if (CheckStatus(inputStatus.Id) == 1)
-                            {
-                                break;
-                            }
-                        }
-                        catch (FalkonryClient.Service.FalkonryException e)
-                        {
-                            log.Error(e.GetBaseException() + "\n");
-                            Console.ReadKey();
-                            Environment.Exit(1);
-                        }
-                        catch (Exception ex)
-                        {
-                            log.WarnFormat(ex.Message + "\n" + "Retry Attempt: {0}", i + 1);
-                        }
-                    }
-                    if (i == 3)
-                    {
-                        log.Error("Cannot add data to the datastream!");
-                        Console.ReadKey();
-                        Environment.Exit(1);
-                    }
-                }
-                else
-                {
-                    try
-                    {
-                        inputStatus = falkonry.AddInputStream(datastreamId, stream, options);
-                    }
-                    catch (FalkonryClient.Service.FalkonryException e)
-                    {
-                        log.Error(e.GetBaseException() + "\n");
-                    }
-                }
-            }
-        }
-
-        public void GetLiveOutput(String assessmentId)
-        {
-            void EventSource_Message(object sender, EventSource.ServerSentEventArgs e)
-            {
-                try
-                {
-                    var falkonryEvent = JsonConvert.DeserializeObject<FalkonryEvent>(e.Data);
-                    log.Info(falkonryEvent.ToString() + "\n");
-                }
-                catch (System.Exception exception)
-                {
-                    log.Error(exception.Message);
-                }
-
-            }
-
-            //On any error while getting live streaming output, EventSource_Error will be triggered
-            void EventSource_Error(object sender, EventSource.ServerSentErrorEventArgs e)
-            {
-                Console.Write(e.Exception.Message + "\n");
-                Console.Write(e.Exception.StackTrace + "\n");
-            }
-
-            EventSource eventSource = falkonry.GetOutput(assessmentId, null, null);
-            eventSource.Message += EventSource_Message;
-            eventSource.Error += EventSource_Error;
-
-            log.Info("Press any key to stop listening.\n");
-            Console.ReadKey();
-            //eventSource.Dispose();
-        }
-
-        private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-
         public static void Main(String[] args)
         {
             /*
@@ -326,15 +31,19 @@ namespace ConsoleApp2
                 to the getData() method of the file adapter.
             */
 
-            //FileAdapter fileAdapter = new FileAdapter();
-            //ADKConnnector aDK = new ADKConnnector();
+            //FileAdapter f = new FileAdapter();
+            //ADKConnHist aDK = new ADKConnHist();
 
             //var datastreamId = aDK.CreateDataStream();
 
             //String fileName = "fileName";
-            //var stream = fileAdapter.GetData(fileName);
+            //var stream = f.GetData(fileName);
             //String fileType = "csv";
-            //aDK.PostData(datastreamId, stream, fileType, liveStatus=false);
+
+            ////  TODO: Uncomment the following line for json file.
+            //// fileType = "json";
+
+            //aDK.IngestData(datastreamId, stream, fileType);
 
             //  ###############################################################################
 
@@ -345,15 +54,20 @@ namespace ConsoleApp2
                 getDataStream() method of the file adapter.
             */
 
-            //FileAdapter fileAdapter = new FileAdapter();
-            //ADKConnnector aDK = new ADKConnnector();
+            //FileAdapter f = new FileAdapter();
+            //ADKConnHist aDK = new ADKConnHist();
 
             //var datastreamId = aDK.CreateDataStream();
 
             //String fileName = "fileName";
-            //var stream = fileAdapter.GetStream(fileName);
+            //var stream = f.GetStream(fileName);
             //String fileType = "csv";
-            //aDK.PostDataFromStream(datastreamId, stream, fileType, liveStatus = false);
+
+            ////  TODO: Uncomment the following line for json file.
+            //// fileType = "json";
+
+            //aDK.IngestDataFromFile(datastreamId, stream, fileType);
+
 
             //  #############################################################################################
 
@@ -365,33 +79,32 @@ namespace ConsoleApp2
                 NOTE:-
                  1. Go on the Falkonry UI and build a model.
                  2. After building a model, run the code.
-                 While using this method you can choose between postData(), postDataFromStream() and postMoreDataFromStream()
+                 While using this method you can choose between IngestData(), IngestDataFromFile() and IngestDataFromFolder()
                  and set live parameter as true.
             */
 
-            //FileAdapter fileAdapter = new FileAdapter();
-            //ADKConnnector aDK = new ADKConnnector();
+            //FileAdapter f = new FileAdapter();
+            //ADKConnLive aDK = new ADKConnLive();
 
             //String datastreamId = "datastreamId";
             //String assessmentId = "assessmentId";
 
-            //aDK.TurnOnLiveMonitoring(datastreamId);
-
             //String fileName = "fileName";
-            //var stream = fileAdapter.GetStream(fileName);
+            //var stream = f.GetStream(fileName);
             //String fileType = "csv";
-            //aDK.PostDataFromStream(datastreamId, stream, fileType, liveStatus = true);
+
+            ////  TODO: Uncomment the following line for json file.
+            //// fileType = "json";
 
             //Parallel.Invoke(() =>
             //{
-            //    aDK.PostDataFromStream(datastreamId, stream, fileType, liveStatus = true);
+            //    aDK.IngestDataFromFile(datastreamId, stream, fileType);
             //},
             //() =>
             //{
             //    aDK.GetLiveOutput(assessmentId);
             //});
 
-            //aDK.TurnOffLiveMonitoring(datastreamId);
 
             //  #########################################################################
 
@@ -402,11 +115,65 @@ namespace ConsoleApp2
                 the folder containing multiple files to the datastream. You will have to give the folder path and pass it to the
                 addMoreHistoricalDataFromStream() in the adk connector
             */
-            //ADKConnnector aDK = new ADKConnnector();
+            //ADKConnHist aDK = new ADKConnHist();
 
             //String folderPath = "folderPath";
-            //String datastreamId = "datastreamId";
-            //aDK.PostMoreDataFromStream(datastreamId, folderPath, liveStatus = false);
+            //String datastreamId = aDK.CreateDataStream();
+            //aDK.IngestDataFromFolder(datastreamId, folderPath);
+
+            //  ########################################################################################################
+
+            /*
+                #################### For adding facts data to an existing assessment.  ################
+                The code below will add facts to an existing assessment in the form of a string.
+                And for adding facts the model must be trained by the Falkonry UI.
+            */
+
+            //ADKConnHist aDK = new ADKConnHist();
+            //FileAdapter f = new FileAdapter();
+            //String fileName = "fileName";
+            //String fileType = "csv";
+            //String datastreamId = "pdw7y6hck6k8vw";
+            //String assessmentId = "k7vmjmlqqhtdtt";
+
+            ////  TODO: Uncomment the following line for json file.
+            //// fileType = "json";
+
+            //String stream = f.GetData(fileName);
+            //aDK.IngestFactsData(datastreamId, assessmentId, stream, fileType);
+
+            //  ########################################################################################################
+
+            /*
+                #################### For adding facts data from a stream to an existing assessment.################
+                The code below will add facts to an existing assessment in the form of a stream.
+                And for adding facts the model must be trained by the Merlin.
+            */
+
+            //ADKConnHist aDK = new ADKConnHist();
+            //FileAdapter f = new FileAdapter();
+            //String fileName = "fileName";
+            //String fileType = "csv";
+            //String datastreamId = "datastreamID";
+            //String assessmentId = "assessmentID";
+
+            ////  TODO: Uncomment the following line for json file.
+            //// fileType = "json";
+
+            //byte[] stream = f.GetStream(fileName);
+            //aDK.IngestFactsFromFile(datastreamId, assessmentId, stream, fileType);
+
+            //  ########################################################################################################
+
+            /*
+                #################### Exporting Facts ################
+                This method will export the existing facts and then store them in a file named exportedFacts.json in the resources
+                folder.
+            */
+            //ADKConnHist aDK = new ADKConnHist();
+            //String assessmentId = "k7vmjmlqqhtdtt";
+            //String datastreamId = "pdw7y6hck6k8vw";
+            //aDK.ExportFacts(datastreamId, assessmentId);
 
             //  ########################################################################################################
